requirements.txt
langchain
python-dotenv
ipykernel
pypdf
langchain-community
beautifulsoup4
arxiv
pymupdf
langchain-text-splitters
sentence-transformers
langchain_huggingface
chromadb
langchain_chroma
langchain_groq
langchain_core
fastapi
uvicorn
langserve
streamlit
faiss-cpu
arxiv
wikipedia
duckduckgo-search
youtube_transcript_api
pytube
unstructured
huggingface_hub
gradio
openai
bs4
langchain-openai
sentence_transformers
duckdb
pandas
langchain-groq
duckduckgo_search
mysql-connector-python
SQLAlchemy
validators
numexpr
neo4j
langchain_google_community
cassio
langgraph